RPC improvements (chain id, config options)
