+++ src/Main.hs
@@ -1,16 +1,13 @@
 {-# LANGUAGE RecordWildCards #-}
-{-# LANGUAGE DeriveAnyClass #-}
 
 module Main where
 
 import Control.Lens (view)
-import Control.Monad (unless, forM_)
+import Control.Monad (unless, forM_, when)
 import Control.Monad.Reader (runReaderT)
 import Control.Monad.Random (getRandomR)
-import Data.Aeson (ToJSON, FromJSON, ToJSONKey)
 import Data.Aeson qualified as JSON
 import Data.Aeson.Key qualified as Aeson.Key
-import Data.Aeson.Types (toJSONKeyText)
 import Data.ByteString qualified as BS
 import Data.ByteString.UTF8 qualified as UTF8
 import Data.Function ((&))
@@ -26,19 +23,16 @@ import Data.Text qualified as Text
 import Data.Time.Clock.System (getSystemTime, systemSeconds)
 import Data.Vector qualified as Vector
 import Data.Version (showVersion)
-import GHC.Generics (Generic)
 import Main.Utf8 (withUtf8)
 import Options.Applicative
 import Paths_echidna (version)
 import System.Directory (createDirectoryIfMissing, doesFileExist)
-import System.Environment (lookupEnv)
 import System.Exit (exitWith, exitSuccess, ExitCode(..))
 import System.FilePath ((</>))
 import System.IO (hPutStrLn, stderr)
 import System.IO.CodePage (withCP65001)
-import Text.Read (readMaybe)
 
-import EVM (Contract(..), bytecode, ContractCode (RuntimeCode), RuntimeCode (ConcreteRuntimeCode), initialContract)
+import EVM (Contract(..), bytecode)
 import EVM.Dapp (dappInfo)
 import EVM.Solidity (SolcContract(..), SourceCache(..))
 import EVM.Types (Addr, keccak', W256)
@@ -55,6 +49,7 @@ import Echidna.Campaign (isSuccessful)
 import Echidna.UI
 import Echidna.Output.Source
 import Echidna.Output.Corpus
+import Echidna.RPC qualified as RPC
 import Echidna.Solidity (compileContracts, selectSourceCache)
 import Echidna.Utility (measureIO)
 import Etherscan qualified
@@ -62,31 +57,30 @@ import Etherscan qualified
 main :: IO ()
 main = withUtf8 $ withCP65001 $ do
   opts@Options{..} <- execParser optsParser
-  seed <- getRandomR (0, maxBound)
   EConfigWithUsage loadedCfg ks _ <-
     maybe (pure (EConfigWithUsage defaultConfig mempty mempty)) parseConfig cliConfigFilepath
-  let cfg = overrideConfig loadedCfg opts
+  cfg <- overrideConfig loadedCfg opts
+
   unless cfg.solConf.quiet $
     forM_ ks $ hPutStrLn stderr . ("Warning: unused option: " ++) . Aeson.Key.toString
 
   -- Try to load the persisted RPC cache. TODO: we use the corpus dir for now,
   -- think where to place it
-  maybeBlock :: Maybe Int <- lookupEnv "ECHIDNA_RPC_BLOCK" <&> (>>= readMaybe)
   (loadedContractsCache, loadedSlotsCache) <-
     case cfg.campaignConf.corpusDir of
       Nothing -> pure (Nothing, Nothing)
       Just dir -> do
         let cache_dir = dir </> "cache"
         createDirectoryIfMissing True cache_dir
-        case maybeBlock of
+        case cfg.rpcBlock of
           Just block -> do
-            parsedContracts :: Maybe (Map Addr FetchedContractData) <-
+            parsedContracts :: Maybe (Map Addr RPC.FetchedContractData) <-
               readFileIfExists (cache_dir </> "block_" <> show block <> "_fetch_cache_contracts.json")
               <&> (>>= JSON.decodeStrict)
             parsedSlots :: Maybe (Map Addr (Map W256 (Maybe W256))) <-
               readFileIfExists (cache_dir </> "block_" <> show block <> "_fetch_cache_slots.json")
               <&> (>>= JSON.decodeStrict)
-            pure (Map.map (Just . fromFetchedContractData) <$> parsedContracts, parsedSlots)
+            pure (Map.map (Just . RPC.fromFetchedContractData) <$> parsedContracts, parsedSlots)
           Nothing ->
             pure (Nothing, Nothing)
 
@@ -97,13 +91,18 @@ main = withUtf8 $ withCP65001 $ do
   cacheContractsRef <- newIORef $ fromMaybe mempty loadedContractsCache
   cacheSlotsRef <- newIORef $ fromMaybe mempty loadedSlotsCache
   cacheMetaRef <- newIORef mempty
+  chainId <- RPC.fetchChainId cfg.rpcUrl
+
   let env = Env { cfg = cfg
                   -- TODO put in real path
                 , dapp = dappInfo "/" solcByName sourceCache
                 , metadataCache = cacheMetaRef
                 , fetchContractCache = cacheContractsRef
-                , fetchSlotCache = cacheSlotsRef }
+                , fetchSlotCache = cacheSlotsRef
+                , chainId = chainId
+                }
 
+  seed <- getRandomR (0, maxBound)
   (vm, world, echidnaTests, dict) <- prepareContract env contracts cliFilePath cliSelectedContract seed
 
   initialCorpus <- loadInitialCorpus env world
@@ -118,11 +117,11 @@ main = withUtf8 $ withCP65001 $ do
     Nothing -> pure ()
     Just dir -> do
       let cache_dir = dir </> "cache"
-      case maybeBlock of
+      case cfg.rpcBlock of
         Just block -> do
           -- Save fetched data, it's okay to override as the cache only grows
           JSON.encodeFile (cache_dir </> "block_" <> show block <> "_fetch_cache_contracts.json")
-                          (toFetchedContractData <$> Map.mapMaybe id contractsCache)
+                          (RPC.toFetchedContractData <$> Map.mapMaybe id contractsCache)
           JSON.encodeFile (cache_dir </> "block_" <> show block <> "_fetch_cache_slots.json")
                           slotsCache
         Nothing ->
@@ -133,28 +132,31 @@ main = withUtf8 $ withCP65001 $ do
       measureIO cfg.solConf.quiet "Saving corpus" $
         saveTxs (dir </> "coverage") (snd <$> Set.toList campaign.corpus)
 
-      -- TODO: Add another option to config for saving coverage report
+      -- TODO: We use the corpus dir to save coverage reports which is confusing.
+      -- Add config option to pass dir for saving coverage report and decouple it
+      -- from corpusDir.
+      when cfg.campaignConf.coverageReport $ do
+        -- We need runId to have a unique directory to save files under so they
+        -- don't collide with the next runs. We use the current time for this
+        -- as it orders the runs chronologically.
+        runId <- fromIntegral . systemSeconds <$> getSystemTime
 
-      -- We need runId to have a unique directory to save files under so they
-      -- don't collide with the next runs. We use the current time for this
-      -- as it orders the runs chronologically.
-      runId <- fromIntegral . systemSeconds <$> getSystemTime
+        -- coverage reports for external contracts
+        forM_ (Map.toList contractsCache) $ \(addr, mc) ->
+          case mc of
+            Just contract -> do
+              r <- externalSolcContract addr contract
+              case r of
+                Just (externalSourceCache, solcContract) -> do
+                  let dir' = dir </> show addr
+                  saveCoverage False runId dir' externalSourceCache [solcContract] campaign.coverage
+                  saveCoverage True  runId dir' externalSourceCache [solcContract] campaign.coverage
+                Nothing -> pure ()
+            Nothing -> pure ()
 
-      forM_ (Map.toList contractsCache) $ \(addr, mc) ->
-        case mc of
-          Just contract -> do
-            r <- externalSolcContract addr contract
-            case r of
-              Just (externalSourceCache, solcContract) -> do
-                let dir' = dir </> show addr
-                saveCoverage False runId dir' externalSourceCache [solcContract] campaign.coverage
-                saveCoverage True  runId dir' externalSourceCache [solcContract] campaign.coverage
-              Nothing -> pure ()
-          Nothing -> pure ()
-
-      -- save source coverage reports
-      saveCoverage False runId dir sourceCache contracts campaign.coverage
-      saveCoverage True  runId dir sourceCache contracts campaign.coverage
+        -- save source coverage reports
+        saveCoverage False runId dir sourceCache contracts campaign.coverage
+        saveCoverage True  runId dir sourceCache contracts campaign.coverage
 
   if isSuccessful campaign then exitSuccess else exitWith (ExitFailure 1)
 
@@ -195,35 +197,6 @@ main = withUtf8 $ withCP65001 $ do
             , immutableReferences = mempty
             }
       pure (sourceCache, solcContract)
-
-data FetchedContractData = FetchedContractData
-  { runtimeCode :: BS.ByteString
-  , nonce :: W256
-  , balance :: W256
-  }
-  deriving (Generic, ToJSON, FromJSON, Show)
-
-fromFetchedContractData :: FetchedContractData -> Contract
-fromFetchedContractData contractData =
-  (initialContract (EVM.RuntimeCode (EVM.ConcreteRuntimeCode contractData.runtimeCode)))
-    { _nonce = contractData.nonce
-    , _balance = contractData.balance
-    , _external = True
-    }
-
-toFetchedContractData :: Contract -> FetchedContractData
-toFetchedContractData contract =
-  let code = case contract._contractcode of
-               RuntimeCode (ConcreteRuntimeCode c) -> c
-               _ -> error "unexpected code"
-  in FetchedContractData
-    { runtimeCode = code
-    , nonce = contract._nonce
-    , balance = contract._balance
-    }
-
-instance ToJSONKey W256 where
-  toJSONKey = toJSONKeyText (Text.pack . show)
 
 readFileIfExists :: FilePath -> IO (Maybe BS.ByteString)
 readFileIfExists path = do
@@ -307,12 +280,17 @@ versionOption = infoOption
                   ("Echidna " ++ showVersion version)
                   (long "version" <> help "Show version")
 
-overrideConfig :: EConfig -> Options -> EConfig
-overrideConfig config Options{..} =
-  config { solConf = overrideSolConf config.solConf
-         , campaignConf = overrideCampaignConf config.campaignConf
-         }
-         & overrideFormat
+overrideConfig :: EConfig -> Options -> IO EConfig
+overrideConfig config Options{..} = do
+  rpcUrl <- RPC.rpcUrlEnv
+  rpcBlock <- RPC.rpcBlockEnv
+  pure $
+    config { solConf = overrideSolConf config.solConf
+           , campaignConf = overrideCampaignConf config.campaignConf
+           , rpcUrl = rpcUrl <|> config.rpcUrl
+           , rpcBlock = rpcBlock <|> config.rpcBlock
+           }
+           & overrideFormat
   where
     overrideFormat cfg =
       case maybe cfg.uiConf.operationMode NonInteractive cliOutputFormat of
