+++ src/test/Common.hs
@@ -103,7 +103,8 @@ runContract f selectedContract cfg = do
                 , dapp = dappInfo "/" solcByName sourceCache
                 , metadataCache = cacheMeta
                 , fetchContractCache = cacheContracts
-                , fetchSlotCache = cacheSlots }
+                , fetchSlotCache = cacheSlots
+                , chainId = Nothing }
   (vm, world, echidnaTests, dict) <- prepareContract env contracts (f :| []) selectedContract seed
   let corpus = []
   -- start ui and run tests
@@ -136,7 +137,8 @@ checkConstructorConditions fp as = testCase fp $ do
                 , dapp = emptyDapp
                 , metadataCache = cacheMeta
                 , fetchContractCache = cacheContracts
-                , fetchSlotCache = cacheSlots }
+                , fetchSlotCache = cacheSlots
+                , chainId = Nothing }
   (v, _, t) <- loadSolTests env (fp :| []) Nothing
   r <- flip runReaderT env $
     mapM (\u -> evalStateT (checkETest u) v) t
