+++ tests/solidity/basic/default.yaml
@@ -81,3 +81,9 @@ corpusDir: null
 mutConsts: [1, 1, 1, 1]
 # maximum value to send to payable functions
 maxValue: 100000000000000000000 # 100 eth
+# URL to fetch contracts over RPC
+rpcUrl: null
+# block number to use when fetching over RPC
+rpcBlock: null
+# whether or not to generate a coverage report
+coverageReport: false
